Create TravelExpenses.java
Write a program to estimate the cost of an upcoming vacation. There are three  types of expenses: food, boarding, and entertainments. Food, boarding, and  entertainments are based on cost per day and estimated days for each one of  them. Test the program with main class by giving you own values and print  the output using jdbc